Here is the Second Version Batch of Old Canned Projects
In this Commit I've added in Source Code of some of my Old Cancelled Versions of my Fan Projects. And added in a Folder for the Compiled Cancelled Old Versions of my Rom Hacks and Reordered files and Added in some Extra files like Save Files and Save States for Sonic Advance 3 Test Hack and more hope you people enjoy the new stuff that was added ^-^ -Edit for some Reason there was a problem with many folders and files inside of the Source Code Folder so had to fix the errors and edit this commit to feature the rest of the Source Code Folders and Files. also this edit feature a whole lot of changes so this commit is huge so yeah lots to look at 😆